refactor: extracted PreRelease to module and create tests
